Added own message queue

+++ src/manager.h
@@ -1,8 +1,6 @@
 #ifndef MANAGER
 #define MANAGER
 
-#include <Arduino.h>
-
 #include "pin.h"
 
 class Manager {
@@ -18,7 +16,7 @@ private:
 public:
   Manager();
   static Manager *getInstance() { return managerInstance; }
-  bool readPin(byte pin, byte *valueOut);
+  bool readPin(byte pin, unsigned int *valueOut);
   bool writePin(byte pin, byte writeValue);
 };
